+++ elements/layercontrol/src/methods/layercontrol/first-update.js
@@ -1,3 +1,5 @@
+import { getElement } from "../../../../../utils";
+
 /**
  * Retrieves the EOxMap element based on the provided selector in EOxLayerControl's 'for' property.
  * Updates the EOxLayerControl's map property based on the target element's map.
@@ -6,11 +8,7 @@
  * @returns {import("@eox/map/main").EOxMap | null} The EOxMap element or null if not found.
  */
 const firstUpdatedMethod = (EOxLayerControl) => {
-  const mapSelector = EOxLayerControl.for;
-
-  // Find the element specified in EOxLayerControl.for
-  /** * @type {Element & { map: import("ol").Map }} */
-  const foundElement = document.querySelector(mapSelector);
+  const foundElement = getElement(EOxLayerControl.for);
 
   // Check if the element exists and its map property differs from EOxLayerControl.map
   if (foundElement && foundElement.map !== EOxLayerControl.map) {