+++ elements/map/main.ts
@@ -40,6 +40,7 @@ import {
   get as getProjection,
 } from "ol/proj";
 import { Coordinate } from "ol/coordinate";
+import { getElement } from "../../utils";
 
 type EOxAnimationOptions = import("ol/View").AnimationOptions &
   import("ol/View").FitOptions;
@@ -213,6 +214,16 @@ export class EOxMap extends LitElement {
           )
         ) {
           const layerToBeRemoved = getLayerById(this, l.properties?.id);
+          const jsonDefinition = layerToBeRemoved.get("_jsonDefinition");
+          jsonDefinition.interactions?.forEach(
+            (interaction: import("./src/generate").EOxInteraction) => {
+              if (interaction.type === "select") {
+                this.removeSelect(interaction.options.id);
+              } else {
+                this.removeInteraction(interaction.options.id);
+              }
+            }
+          );
           this.map.removeLayer(layerToBeRemoved);
         }
       });
@@ -389,11 +400,27 @@ export class EOxMap extends LitElement {
     }
   }
 
+  private _sync: string | EOxMap | undefined;
   /**
-   * Sync map with another map view by providing its query selector
+   * Sync map with another map view by providing its query selector or an eox-map DOM element
    */
   @property()
-  sync: string;
+  set sync(sync: string | EOxMap | undefined) {
+    this._sync = sync;
+    if (sync) {
+      // Wait for next render tick
+      setTimeout(() => {
+        const originMap = getElement(sync) as EOxMap;
+        if (originMap) {
+          this.map.setView(originMap.map.getView());
+        }
+      });
+    }
+  }
+
+  get sync() {
+    return this._sync;
+  }
 
   /**
    * The native OpenLayers map object.
@@ -488,7 +515,7 @@ export class EOxMap extends LitElement {
    * Removes a given EOxSelectInteraction from the map.
    * @param id id of the interaction
    */
-  removeSelect = (id: string) => {
+  removeSelect = (id: string | number) => {
     this.selectInteractions[id].remove();
     delete this.selectInteractions[id];
   };
@@ -532,7 +559,7 @@ export class EOxMap extends LitElement {
 
   parseTextToFeature = (
     text: string,
-    vectorLayer: VectorLayer<import("ol/Feature").default>,
+    vectorLayer: VectorLayer,
     replaceFeatures: boolean = false
   ) => {
     parseText(text, vectorLayer, this, replaceFeatures);
@@ -580,13 +607,6 @@ export class EOxMap extends LitElement {
   }
 
   firstUpdated() {
-    if (this.sync) {
-      const originMap: EOxMap = document.querySelector(this.sync);
-      if (originMap) {
-        this.map.setView(originMap.map.getView());
-      }
-    }
-
     this.map.once("change:target", (e) => {
       // set center again after target, as y-coordinate might be 0 otherwise
       e.target.getView().setCenter(this.center);