+++ elements/map/test/vectorTilesLayer.cy.ts
@@ -23,14 +23,12 @@ describe("VectorTile Layer", () => {
     return new Cypress.Promise((resolve) => {
       cy.get("eox-map").should(($el) => {
         const eoxMap = <EOxMap>$el[0];
-        const layer = eoxMap.getLayerById("countries") as VectorTile<
-          import("ol/render/Feature").default
-        >;
+        const layer = eoxMap.getLayerById("countries") as VectorTile;
         setTimeout(() => {
           // to do: not able to wait for rendercomplete directly, as `applyStyle` is async
-          const features = layer
-            .getSource()
-            .getFeaturesInExtent(eoxMap.map.getView().calculateExtent());
+          const features = layer.getFeaturesInExtent(
+            eoxMap.map.getView().calculateExtent()
+          );
           expect(features.length).to.be.greaterThan(10);
           resolve();
         }, 1000);